Function app_dirs::app_root

pub fn app_root(t: AppDataType, app: &AppInfo) -> Result<PathBuf, AppDirsError>

Creates (if necessary) and returns path to app-specific data directory for provided data type.

If the directory structure does not exist, this function will recursively create the full hierarchy. Therefore, a result of Ok guarantees that the returned path exists.