Function app_dirs::get_app_dir

pub fn get_app_dir(t: AppDataType,
                   app: &AppInfo,
                   path: &str)
                   -> Result<PathBuf, AppDirsError>

Returns (but does not create) path to app-specific data subdirectory for provided data type and subdirectory path.

The path parameter should be a valid relative path separated by forward slashes (/).

A result of Ok means that we determined where the data SHOULD go, but it DOES NOT guarantee that the directory actually exists. (See app_dir.)